some docs improved and email confirmation done

+++ src/controllers/auth.controller.js
@@ -1,4 +1,8 @@
-import { createAccessToken, createPasswordToken } from "../libs/jwt.js";
+import {
+  createAccessToken,
+  createEmailToken,
+  createPasswordToken,
+} from "../libs/jwt.js";
 import { RESEND_API_KEY, DOMAIN } from "../config.js";
 import User from "../models/user.model.js";
 //to encrypt the password
@@ -6,6 +10,7 @@ import bcryptjs from "bcryptjs";
 import { uploadImage, deleteImage } from "../utils/cloudinary.js";
 import fs from "fs-extra";
 import { Resend } from "resend";
+import { emailConfirmation } from "../services/resend.js";
 
 export const signUp = async (req, res) => {
   try {
@@ -14,11 +19,11 @@ export const signUp = async (req, res) => {
     const emailFound = await User.findOne({ email });
 
     if (emailFound)
-      return res.status(400).json({ message: "Email already exists" });
+      return res.status(400).json({ message: "Email already in use" });
 
     const usernameFound = await User.findOne({ username });
     if (usernameFound)
-      return res.status(400).json({ message: "Username already exists" });
+      return res.status(400).json({ message: "Username already in use" });
 
     const passHash = await bcryptjs.hash(password, 10);
 
@@ -28,35 +33,55 @@ export const signUp = async (req, res) => {
       username,
     });
 
-    if (req.files?.image) {
-      const result = await uploadImage(req.files.image.tempFilePath);
-      newUser.image = {
-        public_id: result.public_id,
-        secure_url: result.secure_url,
-      };
-      //delete temp files
-      await fs.unlinkSync(req.files.image.tempFilePath);
-    }
-
     const userCreated = await newUser.save();
-    const token = await createAccessToken({
+    const token = await createEmailToken({
       id: userCreated._id,
       superuser: userCreated.superuser,
     });
 
-    res.cookie("token", token);
+    const url = `http://localhost:3000/api/user-confirmed/${userCreated._id}/${token}`;
+
+    await emailConfirmation(email, url);
 
     res.json({
       id: userCreated._id,
       username: userCreated.username,
       email: userCreated.email,
       image: userCreated.image,
-      createdAt: userCreated.createdAt,
     });
-
-    console.log("new user created successfully");
   } catch (error) {
     res.status(500).json({ message: error.message });
+  }
+};
+
+export const userConfirmed = async (req, res) => {
+  try {
+    const newUser = await User.findByIdAndUpdate(req.params.id, {
+      $set: { confirmed: true },
+    });
+    if (!newUser)
+      return res.status(404).json({ message: "User does not exist" });
+    if (req.files?.image) {
+      const result = await uploadImage(req.files.image.tempFilePath);
+      newUser.image = {
+        public_id: result.public_id,
+        secure_url: result.secure_url,
+      };
+      await fs.unlinkSync(req.files.image.tempFilePath);
+    }
+    const token = await createAccessToken({
+      id: newUser._id,
+      superuser: newUser.superuser,
+    });
+    res.cookie("token", token);
+    res.json({
+      id: newUser._id,
+      username: newUser.username,
+      email: newUser.email,
+      image: newUser.image.secure_url,
+    });
+  } catch (error) {
+    return res.status(500).json({ message: error.message });
   }
 };
 
@@ -64,6 +89,11 @@ export const signIn = async (req, res) => {
   const { email, password } = req.body;
   try {
     const userFound = await User.findOne({ email });
+    if (userFound.confirmed === false)
+      return res.status(400).json({
+        message:
+          "email not confirmed, check your inbox for the confirmation link",
+      });
     if (!userFound) return res.status(400).json({ message: "User not found" });
 
     const match = await bcryptjs.compare(password, userFound.password);
@@ -82,7 +112,7 @@ export const signIn = async (req, res) => {
     });
     console.log("Welcome, " + userFound.username);
   } catch (error) {
-    console.log({ message: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
 export const signOut = (req, res) => {
@@ -93,62 +123,76 @@ export const signOut = (req, res) => {
   return res.sendStatus(200);
 };
 
-export const profile = async (req, res) => {
-  const userFound = await User.findById(req.user.id);
+export const myProfile = async (req, res) => {
+  try {
+    const userFound = await User.findById(req.user.id);
 
-  if (!userFound)
-    return res.status(404).json({ message: "User does not exist" });
+    if (!userFound)
+      return res.status(404).json({ message: "User does not exist" });
 
-  return res.json({
-    username: userFound.username,
-    email: userFound.email,
-    createdAt: userFound.createdAt,
-    description: userFound.description,
-    password: userFound.password,
-    image: userFound.image,
-  });
+    return res.json({
+      username: userFound.username,
+      email: userFound.email,
+      createdAt: userFound.createdAt,
+      description: userFound.description,
+      password: userFound.password,
+      image: userFound.image.secure_url,
+    });
+  } catch (error) {
+    return res.status(500).json({ message: error.message });
+  }
+};
+
+export const otherUserProfile = async (req, res) => {
+  try {
+    const userFound = await User.findById(req.params.id);
+
+    if (!userFound)
+      return res.status(404).json({ message: "User does not exist" });
+
+    return res.json({
+      username: userFound.username,
+      description: userFound.description,
+      image: userFound.image.secure_url,
+    });
+  } catch (error) {
+    return res.status(500).json({ message: error.message });
+  }
 };
 
 export const updateProfile = async (req, res) => {
   try {
-    if (req.user.id !== req.params.id) {
-      return res.status(403).json({
-        message: "Unauthorized. You can only modify your own profile.",
-      });
-    } else {
-      const { password, username, email, description } = req.body;
-      const userData = {};
+    const { password, username, email, description } = req.body;
+    const userData = {};
 
-      if (password) {
-        const hashedPassword = await bcryptjs.hash(password, 10);
-        userData.password = hashedPassword;
-      }
-      if (email) userData.email = email;
-
-      if (description) userData.description = description;
-
-      if (username) userData.username = username;
-
-      const userFound = await User.findByIdAndUpdate(req.params.id, userData, {
-        new: true,
-      });
-
-      if (!userFound)
-        return res.status(500).json({ message: "User not found" });
-      if (req.files?.image) {
-        if (userFound.image?.public_id) {
-          await deleteImage(userFound.image.public_id);
-        }
-        const result = await uploadImage(req.files.image.tempFilePath);
-        userFound.image = {
-          public_id: result.public_id,
-          secure_url: result.secure_url,
-        };
-        await fs.unlinkSync(req.files.image.tempFilePath);
-      }
-      res.json(userFound);
-      console.log("Profile updated successfully");
+    if (password) {
+      const hashedPassword = await bcryptjs.hash(password, 10);
+      userData.password = hashedPassword;
     }
+    if (email) userData.email = email;
+
+    if (description) userData.description = description;
+
+    if (username) userData.username = username;
+
+    const userFound = await User.findByIdAndUpdate(req.user.id, userData, {
+      new: true,
+    });
+
+    if (!userFound) return res.status(500).json({ message: "User not found" });
+    if (req.files?.image) {
+      if (userFound.image?.public_id) {
+        await deleteImage(userFound.image.public_id);
+      }
+      const result = await uploadImage(req.files.image.tempFilePath);
+      userFound.image = {
+        public_id: result.public_id,
+        secure_url: result.secure_url,
+      };
+      await fs.unlinkSync(req.files.image.tempFilePath);
+    }
+    res.json(userFound);
+    console.log("Profile updated successfully");
   } catch (error) {
     return res.status(404).json({ message: error.message });
   }
@@ -182,10 +226,14 @@ export const forgotPassword = async (req, res) => {
 export const newPassword = async (req, res) => {
   try {
     const { password } = req.body;
-    const userFound = await User.findById(req.params.id);
-    if (!userFound) return res.status(404).json({ message: "User not found" });
     const passHash = await bcryptjs.hash(password, 10);
-    await User.updateOne({ id: id }, { $set: { password: passHash } });
+    const userFound = await User.findByIdAndUpdate(
+      req.params.id,
+      { $set: { password: passHash } },
+      { new: true }
+    );
+    if (!userFound) return res.status(404).json({ message: "User not found" });
+
     res.json({
       id: userFound._id,
       username: userFound.username,
